Fix compiler warnings
size_t is not an int on Windows 64

+++ Wrapping/vtkParse.tab.c
@@ -183,8 +183,8 @@ int  HaveComment;
 char CommentText[50000];
 int CommentState;
 int sigClosed;
-int sigMark[10];
-int sigMarkDepth = 0;
+size_t sigMark[10];
+size_t sigMarkDepth = 0;
 unsigned int sigAllocatedLength;
 int mainClass;
 char *currentId = 0;
@@ -245,7 +245,7 @@ void preSig(const char *arg)
     }
   else if (!sigClosed)
     {
-    int m, n;
+    size_t m, n;
     char *cp;
     checkSigSize(arg);
     cp = currentFunction->Signature;
@@ -267,7 +267,7 @@ void postSig(const char *arg)
     }
   else if (!sigClosed)
     {
-    int m, n;
+    size_t m, n;
     char *cp;
     checkSigSize(arg);
     cp = currentFunction->Signature;
@@ -288,7 +288,7 @@ void preScopeSig(const char *arg)
     }
   else if (!sigClosed)
     {
-    int i, m, n, depth;
+    size_t i, m, n, depth;
     char *cp;
     checkSigSize(arg);
     cp = currentFunction->Signature;
@@ -365,7 +365,7 @@ void swapSig()
     }
   if (currentFunction->Signature && sigMark[sigMarkDepth] > 0)
     {
-    int i, m, n, nn;
+    size_t i, m, n, nn;
     char c;
     char *cp;
     cp = currentFunction->Signature;
@@ -394,7 +394,7 @@ void chopSig(void)
 {
   if (currentFunction->Signature)
     {
-    int n = strlen(currentFunction->Signature);
+    size_t n = strlen(currentFunction->Signature);
     if (n > 0)
       {
       currentFunction->Signature[n-1] = '\0';
